Pushing for first release :ribbon:
working a lot on the admin panel, started the header stuff, will get
working on a cool calendar edit page, after that I want to work on image
management. After that, it's time to set up a test enviroment.

For full release, need to be able to edit noticebar bit, calendar and
images. Can work on more functionality afterwards, esp for the sixth
form stuff.