working on the first script again
Basically.... I'm really not happy with the first script being hardcoded
to specific pi names. Although I might hardcode them in the beginning,
what I really want to do is dynamically find all the Pi's that have been
deployed then deploy to them as an when is needed.

Thought: what about the pi's announce their existance to the server, the
server then adds the pi's name to a list.txt, then the script get's all
the pi names from the list and does for loops and stuff to do what needs
to be done.

if a pi gets taken out of service, remove it from the list.

:thought_balloon: